+++ src/pydbm/database/__init__.py
@@ -1,0 +1,3 @@
+from pydbm.database.manager import DatabaseManager
+
+__all__ = ("DatabaseManager",)