+++ src/pydbm/models/fields/base.py
@@ -10,13 +10,16 @@ from pydbm.models import validators as odbm_validators
 if typing.TYPE_CHECKING:
     from pydbm.models.base import BaseModel
     from pydbm.models.meta import Meta
+    from pydbm.typing_extra import AnyCallableFunctionT, ValidatorsT
 
 
-__all__ = ["AnyCallableFunctionT", "BaseField", "Field", "Undefined", "ValidatorsT"]
+__all__ = (
+    "BaseField",
+    "Field",
+    "Undefined",
+)
 
 
-AnyCallableFunctionT = typing.List[typing.Callable[[typing.Any], typing.Any]]
-ValidatorsT = typing.List[typing.Callable[[typing.Any], typing.Optional[bool]]]
 Self = typing.TypeVar("Self", bound="BaseField")  # unexport: not-public
 
 Undefined = type("Undefined", (), {"__repr__": lambda self: "Undefined", "__name__": "Undefined"})()
@@ -37,18 +40,17 @@ class BaseField:
         "max_value",
         "min_value",
         "kwargs",
-        "__is_call_run",
+        "_is_call_run",
     )
 
     def __init__(
         self,
         default: typing.Any = Undefined,
         default_factory: typing.Callable[[], typing.Any] = Undefined,
-        normalizers: AnyCallableFunctionT | None = None,
-        validators: ValidatorsT | None = None,
+        normalizers: list[AnyCallableFunctionT] | None = None,
+        validators: list[ValidatorsT] | None = None,
         max_value: int | None = None,
         min_value: int | None = None,
-        **kwargs,
     ) -> None:  # noqa: E501
         assert not (
             default is not Undefined and default_factory is not Undefined
@@ -56,12 +58,12 @@ class BaseField:
 
         self.default = default
         self.default_factory = default_factory
-        self.normalizers: AnyCallableFunctionT = [] if normalizers is None else normalizers
-        self.validators: ValidatorsT = [] if validators is None else validators
+        self.normalizers: list[AnyCallableFunctionT] = [] if normalizers is None else normalizers
+        self.validators: list[ValidatorsT] = [] if validators is None else validators
         self.max_value = max_value
         self.min_value = min_value
 
-        self.__is_call_run = False
+        self._is_call_run = False
 
     def __set_name__(self, instance: Meta, name: str) -> None:
         self.public_name = name
@@ -90,7 +92,7 @@ class BaseField:
         setattr(instance, self.private_name, normalize_and_validate_value)
 
     def __call__(self: Self, field_name: str, field_type_name: str, *args, **kwargs) -> Self:  # type: ignore[valid-type]  # noqa: E501
-        self.__is_call_run = True
+        self._is_call_run = True
 
         self.field_name = field_name
         self.field_type_name = field_type_name
@@ -98,11 +100,11 @@ class BaseField:
         self.public_name = field_name
         self.private_name = "_" + field_name
 
-        self.validators.append(getattr(odbm_validators, f"validate_{self.field_type_name}"))
+        self.validators.append(getattr(odbm_validators, f"validate_{self.field_type_name.lower()}"))
         return self
 
     def __repr__(self) -> str:
-        if self.__is_call_run:
+        if self._is_call_run:
             return f"{self.__class__.__name__}({self.field_name!r}, {self.field_type_name!r})"
         return (
             f"{self.__class__.__name__}("